feat(payload): migrate following data to Payload CMS collection
- Add Following collection with title, href, description, cta, category, and sortOrder
- Create query functions (getAllFollowing, getFollowingByCategory) for data fetching
- Convert hardcoded following.ts to following.seed.ts with 44 entries across 3 categories
- Update Following page to async server component consuming Payload queries
- Remove deprecated FollowingEntry and FollowingCategory types from content/data
