+++ lib/payload/collections/index.ts
@@ -1,3 +1,4 @@
+export { Following } from "./Following";
 export { Media } from "./Media";
 export { Projects } from "./Projects";
 export { Roles } from "./Roles";
