+++ lib/payload/data/following.seed.ts
@@ -1,0 +1,380 @@
+import type { FollowingSeed } from "./types";
+
+export const FOLLOWING_SEED: FollowingSeed[] = [
+	// Newsletters + Blogs (9 items)
+	{
+		title: "Josh Comeau",
+		href: "https://www.joshwcomeau.com/",
+		description: "All things CSS",
+		cta: "Subscribe",
+		category: "Newsletters + Blogs",
+		sortOrder: 0,
+	},
+	{
+		title: "Alex Zajac",
+		href: "https://www.hungryminds.dev/",
+		description: "Get smarter about Software and AI in 5 minutes.",
+		cta: "Subscribe",
+		category: "Newsletters + Blogs",
+		sortOrder: 1,
+	},
+	{
+		title: "NextJS Weekly",
+		href: "https://nextjsweekly.com/",
+		description:
+			"A newsletter of the best Next.js articles, tools and projects.",
+		cta: "Subscribe",
+		category: "Newsletters + Blogs",
+		sortOrder: 2,
+	},
+	{
+		title: "Remix Team",
+		href: "https://remix.run/blog",
+		description: "Articles from the team behind React Router and Remix.",
+		cta: "Subscribe",
+		category: "Newsletters + Blogs",
+		sortOrder: 3,
+	},
+	{
+		title: "Console",
+		href: "https://console.dev/",
+		description: "A free weekly email digest of the best tools for developers.",
+		cta: "Subscribe",
+		category: "Newsletters + Blogs",
+		sortOrder: 4,
+	},
+	{
+		title: "Robin Wieruch",
+		href: "https://rwieruch.substack.com/",
+		description:
+			"Freelance Web Developer for React.js, Next.js in TypeScript and JavaScript.",
+		cta: "Subscribe",
+		category: "Newsletters + Blogs",
+		sortOrder: 5,
+	},
+	{
+		title: "This Week in React",
+		href: "https://thisweekinreact.com/",
+		description: "The latest news directly in your inbox.",
+		cta: "Subscribe",
+		category: "Newsletters + Blogs",
+		sortOrder: 6,
+	},
+	{
+		title: "Total TypeScript",
+		href: "https://www.totaltypescript.com/newsletter",
+		description:
+			"A comprehensive production-grade TypeScript training by Matt Pocock.",
+		cta: "Subscribe",
+		category: "Newsletters + Blogs",
+		sortOrder: 7,
+	},
+	{
+		title: "Lydia Hallie",
+		href: "https://www.lydiahallie.com/",
+		description: "Software Engineer and Technical Content Specialist",
+		cta: "Read",
+		category: "Newsletters + Blogs",
+		sortOrder: 8,
+	},
+
+	// Podcasts (9 items)
+	{
+		title: "Syntax FM",
+		href: "https://syntax.fm/",
+		description: "Tasty treats for web developers.",
+		cta: "Listen",
+		category: "Podcasts",
+		sortOrder: 0,
+	},
+	{
+		title: "The Changelog",
+		href: "https://changelog.com/podcast",
+		description:
+			"Software's best weekly news brief, deep technical interviews & talk show.",
+		cta: "Listen",
+		category: "Podcasts",
+		sortOrder: 1,
+	},
+	{
+		title: "Front-end Fire",
+		href: "https://front-end-fire.com/",
+		description:
+			"The weekly show that helps you stay up to date on the latest and greatest in the front-end world.",
+		cta: "Listen",
+		category: "Podcasts",
+		sortOrder: 2,
+	},
+	{
+		title: "React Round Up",
+		href: "https://topenddevs.com/podcasts/react-round-up",
+		description: "Learn the latest in React from top end developers.",
+		cta: "Listen",
+		category: "Podcasts",
+		sortOrder: 3,
+	},
+	{
+		title: "Console DevTools",
+		href: "https://console.dev/podcast",
+		description: "Interviews with interesting people in the devtools space.",
+		cta: "Listen",
+		category: "Podcasts",
+		sortOrder: 4,
+	},
+	{
+		title: "DevTools FM",
+		href: "https://www.devtools.fm/",
+		description:
+			"A podcast about developer tools and the people who make them.",
+		cta: "Listen",
+		category: "Podcasts",
+		sortOrder: 5,
+	},
+	{
+		title: "This Week in Tech",
+		href: "https://twit.tv/shows/this-week-in-tech",
+		description:
+			"This Week in Tech is the top-ranked flagship tech podcast from TWiT.tv.",
+		cta: "Listen",
+		category: "Podcasts",
+		sortOrder: 6,
+	},
+	{
+		title: "Waveform",
+		href: "https://podcasts.voxmedia.com/show/waveform-the-mkbhd-podcast",
+		description: "The MKBHD podcast.",
+		cta: "Listen",
+		category: "Podcasts",
+		sortOrder: 7,
+	},
+	{
+		title: "JS Party",
+		href: "https://changelog.com/jsparty",
+		description: "A weekly celebration of JavaScript and the web.",
+		cta: "Listen",
+		category: "Podcasts",
+		sortOrder: 8,
+	},
+
+	// YouTube (26 items)
+	{
+		title: "Theo",
+		href: "https://www.youtube.com/@t3dotgg",
+		description:
+			"Commentary on the industry and changes in tooling and patterns.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 0,
+	},
+	{
+		title: "The Primeagen",
+		href: "https://www.youtube.com/@ThePrimeTimeagen",
+		description:
+			"This is a place for all the things that are awesome on stream.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 1,
+	},
+	{
+		title: "Jack Herrington",
+		href: "https://www.youtube.com/@jherr",
+		description:
+			"Frontend videos from basic to very advanced; tutorials, technology deep dives.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 2,
+	},
+	{
+		title: "ArjanCodes",
+		href: "https://www.youtube.com/@ArjanCodes",
+		description: "Videos about programming and software design.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 3,
+	},
+	{
+		title: "Awesome Coding",
+		href: "https://www.youtube.com/@awesome-coding",
+		description: "Pragmatic coding videos.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 4,
+	},
+	{
+		title: "Ben Awad",
+		href: "https://www.youtube.com/@bawad",
+		description:
+			"Videos about React, React Native, GraphQL, Typescript, Node.js, PostgreSQL, Python, and all things coding.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 5,
+	},
+	{
+		title: "Codevolution",
+		href: "https://www.youtube.com/@Codevolution",
+		description: "Tutorials on the latest tech in web development.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 6,
+	},
+	{
+		title: "Dave Farley",
+		href: "https://www.youtube.com/@ContinuousDelivery",
+		description:
+			"Explores ideas that help to produce Better Software Faster: Continuous Delivery, DevOps, TDD and Software Engineering.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 7,
+	},
+	{
+		title: "Diary of A CEO",
+		href: "https://www.youtube.com/@TheDiaryOfACEO",
+		description: "Life-changing, long-form interviews on a variety of topics.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 8,
+	},
+	{
+		title: "FireShip",
+		href: "https://www.youtube.com/@Fireship",
+		description:
+			"High-intensity code tutorials and tech news to help you ship your app faster.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 9,
+	},
+	{
+		title: "FreeCodeCamp",
+		href: "https://www.youtube.com/@freecodecamp",
+		description: "Excellent learning resource",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 10,
+	},
+	{
+		title: "Fun Fun Function",
+		href: "https://www.youtube.com/@funfunfunction",
+		description: "MPJ's excellent videos",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 11,
+	},
+	{
+		title: "Kent C Dodds",
+		href: "https://www.youtube.com/@KentCDodds-vids",
+		description:
+			"He helps people make the world better through quality software.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 12,
+	},
+	{
+		title: "Kevin Powell",
+		href: "https://www.youtube.com/@KevinPowell",
+		description:
+			"Helping to learn how to make the web, and make it look good while you're at it.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 13,
+	},
+	{
+		title: "Learn with Jason",
+		href: "https://www.youtube.com/@learnwithjason",
+		description:
+			"Live pair programming with expert web devs, career advice, and coding tutorials.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 14,
+	},
+	{
+		title: "Matt Pocock",
+		href: "https://www.youtube.com/@mattpocockuk",
+		description: "The TypeScript wizard.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 15,
+	},
+	{
+		title: "Matt Williams",
+		href: "https://www.youtube.com/@technovangelist",
+		description: "Informational videos about AI, LLMs and ollama",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 16,
+	},
+	{
+		title: "Git Nation - React Conferences",
+		href: "https://www.youtube.com/@ReactConferences",
+		description: "Presentations from react conferences.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 17,
+	},
+	{
+		title: "Git Nation - JavaScript Conferences",
+		href: "https://www.youtube.com/@JavaScriptConferences",
+		description: "Presentations from JS conferences.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 18,
+	},
+	{
+		title: "Sam Selikoff",
+		href: "https://www.youtube.com/@samselikoff",
+		description: "Quality videos on frontend development.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 19,
+	},
+	{
+		title: "SyntaxFM",
+		href: "https://www.youtube.com/@syntaxfm",
+		description: "Tasty developer treats",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 20,
+	},
+	{
+		title: "Traversy Media",
+		href: "https://www.youtube.com/@TraversyMedia",
+		description: "Web development and programming tutorials",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 21,
+	},
+	{
+		title: "Typecraft",
+		href: "https://www.youtube.com/@typecraft_dev",
+		description: "Videos on developer tooling",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 22,
+	},
+	{
+		title: "Valentin Despa",
+		href: "https://www.youtube.com/@vdespa",
+		description:
+			"Learn more about agile software development using tools such as Postman for API testing, Gitlab CI for supporting your DevOps cycle.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 23,
+	},
+	{
+		title: "Web Dev Cody",
+		href: "https://www.youtube.com/@WebDevCody",
+		description:
+			"A collection of various programming videos, tutorials, projects, talks, or anything software engineering and programming related.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 24,
+	},
+	{
+		title: "Web Dev Simplified",
+		href: "https://www.youtube.com/@WebDevSimplified",
+		description:
+			"Teaching web development skills and techniques in an efficient and practical manner.",
+		cta: "Watch",
+		category: "YouTube",
+		sortOrder: 25,
+	},
+];
