+++ lib/payload/data/index.ts
@@ -1,12 +1,14 @@
 // Types
 
 // Seed data
+export { FOLLOWING_SEED } from "./following.seed";
 export { PROJECTS_SEED } from "./projects.seed";
 export { ROLES_SEED } from "./roles.seed";
 export { SERVICES_SEED } from "./services.seed";
 export { SITE_CONTENT_SEED } from "./site-content.seed";
 export { TESTIMONIALS_SEED } from "./testimonials.seed";
 export type {
+	FollowingSeed,
 	ProjectSeed,
 	RoleSeed,
 	ServiceSeed,
