+++ lib/payload/data/types.ts
@@ -20,5 +20,8 @@ export type UseSeed = RequiredDataFromCollectionSlug<"uses">;
 /** Seed type for services - makes id, createdAt, updatedAt optional */
 export type ServiceSeed = RequiredDataFromCollectionSlug<"services">;
 
+/** Seed type for following - makes id, createdAt, updatedAt optional */
+export type FollowingSeed = RequiredDataFromCollectionSlug<"following">;
+
 /** Seed type for site content global - makes id, createdAt, updatedAt optional */
 export type SiteContentSeed = RequiredDataFromCollection<SiteContent>;
